climate: add reporting of TURN_ON and TURN_OFF support
In HA 2024.2, climate devices are expected to report support of
turn_on and turn_off services.

During the deprecation period until 2025.1, warnings are reported in
the log if these methods are present and the device is not reporting
this support. This change will reduce these messages to just those
devices that actually do not support the methods (using the methods
would throw an exception).

Issue #1579