Tongou QCB2: expose online_state as a switch rather than sensor
From observing the Tuya app's activity, it was seen that despite the
name, this is not just an indication of the connectivity, but is used
to trigger frequent updates of power/current/voltage data when these
are being observed rather than every 10 minutes as the default update
rate on these devices seems to be.

Making this available as a switch allows an automation to regularly
trigger real time updates to keep the data flowing from the device.

Discussion #1483