Setup: rethrow exceptions as ConfigEntryNotReady
For Home Assistant to retry setup, exceptions thrown from
async_entry_setup need to be a specific exception class.

This is required to handle devices that are offline at the time
Home Assistant boots (eg, HA comes up before the WiFi router
after a power outage, or a device gets switched off at the wall, and
turned on after HA is up).

Issue #1683