Translations: fix using translated entity names.
At some unadvertised point, HA removed the deprecation notices for
defaulting to device names, and in the process changed the behaviour.

Most of the internal components that support translations seem to be
using EntityDescriptor.name, so our overriding of name is not well
looked after as a stable API.

Now there is a new function use_device_name, which tries to figure out
by complex logic, but doesn't get the same result as the old inline
code. If we override that function to act as we expect, then it works
but the special meaning of UNDEFINED seems to have been removed except
for the self._attr_name case.