Translations: remove "and select entities" from description
This goes way back to the early versions before HA offered a way for
the user to control which entities are visible and enabled, so we had
to allow secondary entities to be filtered by the user.  This step has
only included name since before any languages other than English were
added.