Conditional mappings do not support default mappings.
Defaults for conditions have to be set one level up in the mapping,
as the conditions only look for matches, not keep track of defaults.
There may be some side-effect to changing this, and setting one level
up is an acceptable workaround.

In read-write situations that need to control which options are set
for a specific value, it is sometimes better to list all conditions
rather than use a default.

Main effect is on climate hvac_action, swing_mode where more than just
two options for vertical and horizontal are available, cover actions
where opened/opening, closed/closing need to distinguished based on
position, and complex logic to handle optional vacuum activate and
pause switches.

Issue #1709