Avoid explicit returning of None where easy.
Explicity returning None is not necessary at the end of functions in
Python, since None is the default return code.
Rearrange some functions to eliminate some returning of None just
before the end as well.

Shorter functions are easier to follow, so this change helps
readability for those accustomed to Python.