Peteme recessed lighting: naming change
Keep branding under products, make top level name generic.

This is a fairly common dps mapping for lights, matching rgbww_lightbulb, but in this case it is using the 0-1000 ranges more common in newer lights that use dps 20 and up for light control. It is likely to match other brands though just because there are so many Tuya lightbulbs on the market.

Also default color_mode when it has not been received.