WDYK 2P 63A energy meter: omit last 2 bytes from Phase B and C masks
The phase decoding was originally implemented blindly based on other devices
in #659. Doc confirming this format were supplied in #1078, but still no logs.

Based on the diagnostics in issue #1717, only phase A appears to use the
documented format with 2 bytes voltage, 3 bytes each for current and power,
and 2 bytes for residual current.  Other phases appear to drop the residual
current, meaning everything else needs to shift by two bytes.