Ovlaim fan: change speed from string to integer
Original report in #749 did not come with logs, but the documentation
listed this as a string. But discussion #1728 suggests that this is not
detected correctly, and the difference between Casafan and this affecting
detection is only the type of speed dp. Other than that the color_temp
is present but disabled on Casafan, and presents and timer have different
values, but these do not affect detection.

There is code from early in development to not be strict about the type of
quoted integers, so making this integer should still work if there is older
firmware that really does use string.

Discussion #1728