IR Moes heatpump: identified more dps
PR #1755 contains detailed info on a Moes IR Thermostat (AC Controller)
that appears compatible (possibly the very same device).
This allows some of the previously unknown dps to be identified, and
some to be exposed as secondary entities, including the raw remote
controller so the device can be used to control additional devices.