Feyree EV charger: convert charging toggle button to a switch
The original report was a bit ambiguous about the use of this, and the
description or one of the comments labelled it as related to RFid card
use, so I made it a button type.

This appears to be wrong, and it is a switch in the Tuya app.

Issue #1796