Zemismart roller shade: positions are not inverted
History: this config was originally made with `invert: true` at dps level,
where the desired effect was not produced. After yaml lint tests were added,
this issue was detected and "fixed" by moving it into a `mapping`.  However,
the original presence of that `invert: true` seems to have been a mistake, the
values should not be inverted.

Discussion #1830