Config: don't default IP address to Auto
Although Auto works when localtuya is not installed, it causes a log
message telling the user to report a bug, due to the underlying
tinytuya library not being async.

This is annoying, and explicitly setting an IP address handles dropped
connection recovery better, so try to avoid steering users towards using
Auto by not defaulting to it.

Issue #407