HA 2024.5 workaround attempt 2
Revert previous ineffective workaround, instead opt out of importing
on executor.

The original issue seems to have been caused by changes in HA 2024.4
to initialise integrations on an executor thread instead of in the
event loop.  So when the receive thread is created, it is not being
created from the event loop, and later gets killed along with the
startup cleanup (which we previously carefully avoided).

Opt out of this change for now until it is clearer how to adapt to
the new HA architecture changes.

Issue #1871, #1872