Add Tuya cloud assisted setup flow
Pick up the streamlined Tuya cloud login from the official HA cloud Tuya integration. This allows you to get a code from the Smart Life app and then use the app to login to Tuya with a QR code. All available devices for that account can then be select to add locally. Only the IP address still needs to be discovered so a local scan is performed to get that. Just like the final device connection the scan for the IP can fail if other things (including the Smart Life app just used for login!) are connecting to the device when the scan runs so those should be closed before that and the final connection step.

New flow:

Starting page in the config flow allows users to choose to go down the cloud assisted configuration path or the original manual entry path:
Now you need to provide the 'User Code' available from the Smart Life app:
Then you will get presented with a QR code you scan in the Smart Life app:
Once you scan and approve the login in the app, a list of available devices that the cloud account knows about is shown. Now you must select the device you want to add and, if necessary, the gateway it connects through:
Next the local IP address of the device is found by discovery:
Finally the retrieved data is prepopulated into the existing register local device page:

Rebase of PR #1881