Mark async_actually_start as a callback.
async_actually_start was originally written to execute in the event
loop.  Since HA was modified to automatically run non-decorated
coroutines in the executor thread, this has been running in the wrong
thread, making some of its calls non-thread safe.

- Revert previous fix for #1871, #1872 (PR #1876), as it is not needed
when actually_start is running in the event loop, and has a side
effect of causing #1888 (a separate fix for that has been proposed on
the issue, but this revert also avoids that issue).

Issue #1917