Hydrotherm Dynamic/X8: add optional current temperature dp
An Aquatech Dynamic/X8 branded device appears to be compatible, but reports
the water temperature in dp 16, which was not present on the original
Hydrotherm branded device (it may in fact be gen 7). The inlet temperature
is reported as the cold water temperature at the inlet, as expected, and not
the hot water temperature as PR #1686 observed.  Both devices report outlet
temperature as -50, so that appears to be unused.

To handle these together, an optional dp 16 was added as the
current_temperature, which redirects to inlet_temperature if it is missing.

Issue #1922