Loratap Relay and Zigbee curtain: remove invalid scale of 0
Some PRs slipped through with the `scale: 0` copied from Tuya cloud
docs.

This integration uses a multiplier for scale, not a decimal place
shift like the Tuya cloud uses. This allows temperature scales of 2
that some thermostats use to represent 0.5 degree steps to be
represented.

In any case, unscaled values do not need to include a scale mapping.

Issue noticed in the log attached to issue #1930, though this is not
the device being reported.