Entity naming: use device name if translation_key is device class
translation_key uses device class as a fallback (to allow overriding
names/icons based on classes in the translation files).  But this means
we falsely detect unnamed switches with a device class as having a name.
Check if the translation_key is the same as device_class before telling
HA not to use the default name.

Issue #1856