Temperature unit: return strings from unit_from_ascii
This helper function is just for translating simple ASCII into more
complex Unicode units (C to °C, m3 to m³ etc) to make editing files
easier.

Now that HA converted most of the unit constants into Enum's, they are
converting to instances of enums, which might no work in some cases.

Issue #1855

Unit tests appear to have coverage of this, and were not deteecting
any issues, but this is the only possiblity I can see for the issue
not being resolved.