Moebot S: make "Problem" sensor optional, defaulting to "OK"
Although this seems to have some non-problematic statuses, it is
reportedly stuck on "Tilted", which may be because it is no longer
being sent, and that was the last value. Treat it as non-persistent,
with a default value of OK, to clear non-current problems if that is
the case.

Issue #1983