Remove failing test.
It is becoming more difficult to provide test coverage for the HA config
flow, as undocumented changes keep being made to the preconditions needed
to get things to work properly.  Running under HA, all these conditions are
filled, but in unit tests we need to know what exactly to set up before the
test can run, since we do not have a fully running HA system in place.