Add migrations for entities renamed in this release.
In testing this, a divide-by-zero error was found in the unit tests,
which could potentially have come from a config missing required dps.
A new test was added for this, and for convenience, a new all_entities
method added to the device_config helper to return primary and
secondary entities together to avoid needing to process them
separately. Other tests were updated to use this too.

Some unmigrated "countdown" timers were also found and fixed in this
migration.