config_flow: do not filter out devices from cloud list
Instead of filtering out known devices from the cloud list, filter them from
the list that is diplayed for selecting the new device.

This is because hubs can be already existing but still be a valid choice for a hub.

This issue was made worse by the recent change to cache parent devices for reuse
by other subdevices, as once cached the hub was being filtered out from future use.

Issue #1054