Development: add util to detect possible duplicate config files.
Due to the large number of configs, there is an increasing likelihood
that a config already exists, or a close enough one can be easily modified.
When logs are provided this can be easily checked, but for pull requests
and submitted config files without any other info, this is harder, as
the dps need to be synthsised for matching.

The new util/duplicates.py script synthesises representative dps itself
so the check can be done more easily.

Current implementation is simplistic, just checking dps types like the
config matching of the integration, so relies on manual checking to confirm
whether there is really a match there. Any >50% match will be listed, as
older configs may be missing some dps.

In future the checking could be expanded to compare ranges and mapping
values, and entity types to narrow it down to the point where reported
matches are almost certainly duplicates, so this can be used in CI for pull
requests.