Tests: remove unused imports from translations test
This used to have old code related to checking that entities all had translations
for the config flow. But the step was removed some time ago, so this has been disabled.
After last update, we are now checking that all the strings present in English have
translations to all other languages instead.