device config: handle feature flag dps in entities and mapping lists
Allow feature flag dps to be used to hide mappings and disable
entities.

For entities, this also affects their default hidden state, but this
may depend on startup timing, as HA likely checks this early during
startup, when the device communication may not be available yet. At
least though they will show as Unavailable and the user can hide them,
rather than having a false value that misleads as to that feature working.

Issue #2337