Powerworld PW58410: new config for primarily HVAC use
Due to this device combining Hot water and HVAC controls into a single
"mode" it needs to expose a subset of modes to the standard climate hvac_mode
to have a one to one mapping.

In the original config, there was an assumption that Hot Water would
be permanently on, so the modes exposed in the climate entity were the
combination "hot water+"heat/cool modes.

In the case where the unit is installed as HVAC only, this is not
appropriate so a different config is needed.

Issue #1530