New device form: provide more hand-holding
Quality of submissions seems to have slipped again, after improving
over the past year. There seems to be an influx of new users, so
possibly another influencer has done a video.

The problem this time seems to be that the new users don't know their
way around Home Assistant, and no longer need access to iot.tuya.com
to get a local key, so generally don't have it.