config flow: import json to reformat cloud response
Because the QueryThingsDataModel comes through as a string attribute
on the json response rather than actual json data, we need to parse
and reformat it to have it readable.

Actually we probably want to condense it with custom formatting, but as a
first step use the json features in Python.

Issue #2419