Add utf16b64 type.
This type represents a UTF-16-BE string encoded in base64.
This encoding is used by some devices "raw" encoding type to pass
textual messages often containing Chinese characters or punctuation,
or user supplied messages that may be in any language.
I guess there is a fear that utf-8 may sometimes encode to problematic
bytes, or when they test with Chinese strings, this encoding is more
efficient than utf-8.

To support the encoding found in PR #2388