Create vta+_fan_heater
This creates a new device that wasn't recognised and i tested fully via Studio Code Server. I took inspiration from the nedis_ptc_fan_heater.yaml, but that device has much more functions and dps not marked as optionals, so it wouldn't recognise my device. That Fan heater has a great feature that use constraints the hvac_mode with preset_mode, so the fan only option that is reported as preset, now works as a hvac_mode. I also improved the code by adding a "null" dps_val on the conditions, using the documentation of the integration that previously reported an error when selecting this "custom" hvac_mode. I found an error on the nedis_ptc .yaml timer values, that i intend to correct in another pull request.