Qlima D820A: move fault_code to problem sensor.
This belongs with the problem sensor.
Previously I added these to the primary entity, as I did not think it
would work to have the same dp id twice on the same entity, but it turns
out that works fine, so it is better to put the attribute with the
related entity.