Unknown 2-way meter: use total as main energy reading
- use total as main energy reading, and daily as attribute.
Original used daily as main because that was the first reading encountered.

Portion of PR #2481 which does not involve entity renaming.
The more disruptive naming changes will be included when translations for
those sensors are added as part of issue #1708 (actually since it involves
translation parameters, as part of a new issue which is not filed yet)