hvac_action: simplify logic
Climate now handles the off case in code, so it is not necessary for
the config to handle the often complex logic of checking hvac_mode for
the off state.

Making this change in bulk lets the existing configs serve as a better
example for future contributors, even though the change doesn't really
have any effect on the end result.