snt957w_tde_temp: high temp. limit should be optional
This device was working fine with tuya-local as long as it was added to Tuya app. But once I used cloudcut it, suddenly tuya-local could not recognize it anymore (saying that device is unsupported).
Inspection of the yaml revealed that the device broadcasts matching dpids except for dpid 10. This could be because I removed the device from Tuya app and it reset to factory defaults.

TLDR: making this optional makes the device detectable again.