Light: consider target color mode for brightness adjust method
The logic for controlling brightness relied a lot on the order of operations.
In previous changes to better decide when using the v from hsv was appropriate,
that order was changed, so now there are cases where white brightness overrides
hsv brightness, causing some lights to revert straight to white mode when
they are changed to colour.

This adds an optional parameter for passing the target mode, so that can be
taken into account as well.

Issue #2552