config: reimplement mask as documented.
Mask is documented as being at dp level, but was implemented as being
a mapping.

Mapping makes sense from the point of view that this changes the
values, but it adds extra complexity in the case when there are
mappings for the values, since the mask must be listed for every
mapping. If included in the mapping, it looks like it gives extra
flexibility of using a different mask for different values, but I
can't think of a use use for this, nor can I imagine what the expected
behaviour should be for multiple matches or how reverse mappings
could be implemented in that case.

Making this dp level as documented simplifies some configs, and avoids
the possibility of inconsistent masking within a dp.

PR #1383