light: lights with only brightness control should appear off at 0
Since we started using HA's brightness scaling functions, this has not
been working, as those functions explicitly avoid reporting a
brightness of 0, since they are not expected to be used when the light
is off.  Change the condition to avoid scaling when the value is 0.

Add a test case for this, to catch any future unintentional changes.

Issue #2589