Fix available in mappings to work as intended.
When things get complex, with conditions inside mappings, with
different availability for the conditions and a default in the parent
mapping, the logic was not working.  Move availability up to the point
where we find mappings, but allow mappings with conditions to "slip
through" so the conditions can override the unavailable mapping.

Cleanup on PR #2567