Docs: remove link to Zigbee2MQTT page about tuya debugger.
The debugger only provides links between id and name. It was needed using
the older API Explorer method that only returned dps that fitted the standard
data model to at least get names for other dps that appeared in the log.

The QueryThingsDataModel that is now recommended does not seem to miss any
dps, so the debugger instructions are just a distraction, and have led
users to assume it is an alternative to API Explorer method, when it is not
providing detailed info about the dps, only names.