Beca BHT002: remove device tests
There is nothing unique about this device, the generic device config
tests are adequate to cover it.

Preparation for changes in PR #2640 that will break the tests due to
unknown attribute renaming and moving to another entity it is related
to, also the duplicate floor_temperature in the climate entity removed
since it has it's own entity (which was maybe not originally supported
when this early device config was originally added, then it was left
temporarily for compatibility).