config_flow: protect against cloud info missing some fields
In particular, local_product_key may not be added if local discovery fails,
so this should not be assumed to be available. But it is probably better
to assume everything might be missing, as it has been observed with local_key
in the past, where sub devices do not have their own.

Issue #2643