Avatto blinds: invert direction, modify action logic.
This is a very old config. Since it was added, it became apparent that
most Tuya cover devices have their positions inverted compared with
what HA expects.  This seems to be the cause of #2699 (different
brand, but same product id).

Also lessons learned is to not assume open or closed state when uncertain,
as it blocks some of the UI controls from being used.

Also added problem sensor on dp 10, and moved dp 5 to a select entity (now
that function is understood based on other cover devices).