Eliminate remaining explicit primary and secondary entity references
Initially the integration was designed with a primary entity for each
device, with secondary entities designed to be optionally enabled
according to user preference.

Since then, HA has added entity categories to take care of what is
primary and secondary, which is more flexible, as it allows multiple
primary functions per device, and also made it easier to hide and
disable unwanted entities from the UI, so we dropped the configuration
options for secondary entities, mainly to reduce the effort needed to
maintain translations.

So the distinction between primary entity and secondary entities
doesn't really serve any purpose any more. A few months ago an
all_entities() method was added to simplify all the places that
do the same processing on primary entity and secondary entities.
Locations in the main flow of code were updated to use it, but some
migration, test and utility code was still left using the old methods.

This finishes the task, ready for deprecation of the primary/secondary
distinction.