Don't consider hidden mappings at all when mapping back to dps_val
When a hidden mapping maps to the same dps_val as a normal mapping, the
normal mapping should be used.

But in the previous code, the hidden mapping could be returned from
_find_map_for_value then filtered out later (but only if the hidden
marking was on the map, not an embedded condition).

Instead, do the filtering inside _find_map_for_value, and improve the
logic so that hidden at the map level does not hide all the conditions
within, as that is used in some more complex conditional mappings.

Also take into account available attributes.

PR #2701