Remove the distinction between primary and secondary entities.
This historical distinction has not mattered for many years now.
It is a source of confusion that primary_entity contained one entity
object and secondary_entities contained a list, making the formats
slightly different, creating a barrier to users making configs.

Combine these into a single `entities` list, and deprecate the old
way.