Update ustellar_ut99911_rgb.yaml
Add descriptive values for DP 103 - it denotes the RGB order for the LED strip on the back.

The value can be changed, so this could be a select entity. However I don't see the point of this as it's configured in the factory for the LED strip on the device so users should never change it from the original setting. It will just mess up the colour representation on the LED strip. But it is useful to know the string value to better understand the spec of the LED strip on one of these light bars.