Device config: reduce log verbosity
Even in debug logs we do not need to know that mappings are being "considered",
only what the final result of the mapping is.

Maybe in the early days this was relevant, to debug the mapping logic,
but now it is just noise in the debug log that hides real information.

Inspired by issue #2722