config_flow: log limited dps info when adding devices.
This uses an old API call that does not return all the dps, but it is
better than nothing. Now we can get some dps in the log which will
hopefully help with the lower quality device requests received lately
(though the biggest problem is not including log messages, so
probably not).

On the downside, it may disuade users from looking up and including
the full Query Things Data Model info.