Tests: remove obsolete pending_reset tests.
pending_reset was removed years ago when the integration was changed
to be push by default rather than polled.
The tests continued to work because the initial cached_state with just an
updated_at entry caused has_returned_state to return false.
Since that condition was changed in the last commit to support devices
that do not return any state, the tests now fail. Since they are obsolete,
it is better to remove them than update.