Add support for Zemismart ZMAi-90 smart switch with energy monitor.
Uses a standard smartplug layout, but from the log submitted, the
add_ele dp seems to be acting like a total_forward_energy so it can be
used as a proper meter.

Missing timers, calibration or any extra config options, just the
switch and sensors for energy, current, power, voltage and a fault code.

Issue #2651