Boost score for product_id match to 101%
A product id match signals a perfect match, but at 100% it does not
appear any better than dps that happen to match.

Also do not require all non-optional keys to match when the product id
is already matched. A partial match is still likely to be the correct
device. Only discard the match if there are mismatches in the dps
types.

PR #2773