Add support for text entities.
`text` entities were added to HA in 2022.12, but it went unnoticed.
This completes the set of basic building block entities needed to
expose arbitrary data.

Initially, the RGBCW lightbulb config is using this to expose the
Scene dp so that arbitrary scenes can be set.

This is disabled by default, as it is as advanced feature that
requires some knowledge of how scenes are encoded by the light to use.
The `hidden` attribute at entity level has been added so the entity
could be disabled by default without marking it deprecated or
unavailable.

Issue #2795