Asakuki diffuser: add named_color support
The light on this device had the color disabled due to the format being
inconsistent, and replaced by a separate select entity.

The problem is this made the light white-only from HA point of view,
which should have used a different color mode.

Since HA is breaking backward compatibility with lights that use the color modes
incorrectly in 2025.3, this needs to be fixed by either changing to the
appropriate "brightness" mode instead of "white", or by adding color support
back in. Because we added named color support in the meantime, we are able
to do the latter.

Issue #2807