Use correct Unicode character for DEGREE SIGN
Relative temperatures cannot be labelled with the C and F shortcuts
that translate to the actual unit constants in code, as conversion
would mess them up due to the offset between 0 points of C and F.

So the general degree sign needs to be used instead. It is almost
indistinguishable visually from superscript 0, which was used in
some places due to a convenient keyboard shortcut for inserting it.