Downgrade tinytuya back to 1.15.1
A lot of reports are suddenly coming through where devices are getting no
status returned. Many from devices that were previously working.
I am not seeing it, so it is not universal, and it could also be firmware
updates that are doing it, but it started happening after the last release
that bumped this library, and the changes were quite substantial rework so
there is a possibility that something broke for one of the protocols.

Until we figure out what is happening, or confirm the old version of the
library is also the same, revert to the older version.

There are no features in 1.16.0 that we need, it was just a small step
towards a library that can support async in future.