Entity availability: add option to hide unavailable entities.
The use of available is split between two purposes:

1. To support multiple devices with a single config, based on feature
   flags or availability of optional dps.
2. To dynamically make entities available based on user settable options.

In the first case, we want HA to hide the entities, as they will never
be available for the device. But in the second case, they will
sometimes be available, so we do not want to hide them.

Add a third option for the "hidden" entity attribute which determines
if entities be hidden when "unavailable"

May help with recent reports where users are saying their devices
became unavailable after upgrade, as in most cases we should not be
hiding them just because no data has been received yet (probably HA is
only checking this at startup and will not later enable the entity if
it becomes available).