device communication: ignore HA poll requests once started
We keep our own receive loop to handle push from devices. But when devices
are set to poll only, the receive loop also handles the polling. Handling
the poll requests from HA in this can cause simultaneous requests to be
sent to the device, which many devices cannot handle.

Instead of ignoring the HA requests when a persistent connection is being
used, ignore them when the receive loop is running.

Issue #653, #1804