Add a 3.22 protocol version for enabling "device22" detection
Tinytuya has automatic detection of "device22" devices, which are
3.3 devices that require 3.2 data queries.

This detection seems to falsely detect normal 3.3 devices as device22
if it tries to make a connection when the device is powering up, so
it accepts the network connection but is unable to reply. When this
false detection happens, the tinytuya continues to use device22
commands to communicate with the device, so does not get any data.

With the new 3.22 protocol variant added, we can disable the device22
detection normally, and only enable when the protocol is explicitly 3.22.

Issue #1804