Don't mark devices available if they have not communicated.
A previous change to accept empty polls for devices that don't return data
(such as IR remotes) resulted in devices being marked available even when
they haven't returned any data. But this is not correct when the connection
has not been made at all.

However, we do not want to mark entities as disabled by default based on
the device not having communicated yet on startup, so add an extra condition
to avoid that.