Siguro dehumidifier: add compatible ProBreeze OmniDry 12L
Model is compatible down to the mis-spellings in the mode names.
Seems to report temperature unit and filter reset, but no sign in the
manual that they are actually supported.

Added hidden entities for all the unsupported features except runtime reset,
which makes no sense as there is no runtime and has never been observed.

Issue #2868