Beok TGM50: list products with same id. Expand range to match docs
Issue #1240 and #2989 both give a temperature range of up to 95.
Since the max temperature is configurable, this change can be safely made
without upsetting any existing installation where the range is more limited.

Both the Avatto WT200 and the unbranded M3-GE thermostat have the same
product id and matching dps to the Beok TGM50, so should be the same device.