Smartplugs: disable non-working energy sensors by default
Many smartplugs provide an "add_ele" dp which reports energy usage
over such a short period that it is not usable as an energy sensor
within HA. On some smartplugs this was initially exposed as a sensor
entity in case it was useful in specific cases, but in general we now
know it is not useful, so later configs have this as an attribute of
the switch or power sensor instead.

For those configs that have it as a separate entity, this change
disables the entity by default, to avoid confusing users over its
usefulness. The change also adds the measurement class so that if
users enable it, they can get long term statistics.

Issue #3000