Dongguan garage door opener: change garage to gate.
In evaluating PR #2984, only the class seems to differ significantly.
The PR uses open rather than action, which seems better since the action
is only mapped to opened and closed anyway, and looking at the comments
the original dongguan device seems to also be a gate opener, so class
also updated.