Sefaul Q9 Dehumidifier
Unfortunately this isn't the most intuitive device - neither via the Smartlife App nor the device's control panel.  The device doesn't have a humidity sensor, so the dehumidifier cant natively be triggered that way, and it has a separate air purifier component that gets automatically turned on, so I couldn't really make it work as a dehumidifier the way HA expects.  Perhaps someone with more experience would have a different approach.  But, this config does work the same way as the device's control panel.  

https://www.amazon.com/dp/B0D7959FJX