Morento HY4866-WF: naming adjustments
- Rename filter used to Filter remaining
- Map "medium" air quality to moderate rather than poor

This is based on Morento MR7566-WF PR #3078, which has different presets
and a separate speed control. Likely the "filter used" naming was an assumption
about the way this counted, and is likely the same as the MR7566 as renamed
by a user able to observe how the device actually works.
Air quality good/moderate/severe is more in line with other devices, rather
than good/poor/severe.