Della air conditioner: modify based on PR #3080
This appears close enough to the Sendo Hermes Gold in PR 3080 that it is
almost certainly based on the same control module, perhaps with some model
specific optional behaviour.

Extend the Hermes Gold with the one additional fan mode, and the power limit
entities. Rename "anti frost" to "anti mould", it is likely the original is
also that function (to leave the unit running in dry mode for a while when
it is switched off in high humidity conditions)

Add product details for Sendo device, and rename top level name to be more
generic.