Taxnele energy meter: add additional entities
This was a very early addition, and did not have full info, so it isn't
clear how much (if any) applies to the original device, but an additional
device appears compatible, and on checking the manufacturer (Yueqing Taixin
Electric) the website is taxnele.com. Additional dps are all added as optional.

Issue #2910