Lytmi HDMI sync: rename v component so brightness is preferred
Since the light only supports color mode, brightness is
always sent as the hsv dp, but that interferes with scene selection.
Since the light has a brightness dp, it is obviously supposed to be used,
even thought the light does not support white mode, contrary to the way
other Tuya lights work.

Based on PR #3082