Modify config flow tests to check "special characters" in local key.
There are repeated claims that special characters in local keys do not
work, which have never led to any specific bug being found. Latest
report in issue #3149 gives a specific key that the user claims does
not work, and provides what they thought was a repeatable test case
using dummy data for device id and IP address, but turns out to not be
repeatable.

Use this "known bad" local key in unit tests to make sure there is no
truncation due to any of the characters in the key being processed by
Python that we are missing. Seems to work OK, as expected by the
non-reproducibility of the user's report.