Galaxy projector: add text entities for Theme (Scene) and Music
PR #3119 seems to have a different list of scenes, but with generic names
so not really useful to add as select options. It also defines the same for
Music mode which was not covered in the original config.

Otherwise the config is pretty much the same, except the submitter split
effect out to a separate select, and renamed color_mode to switch (duplicated).